Silence gcc false-positive/format warnings in nii_otsu and overlay check
Two warnings surfaced by the Linux gcc-9 CI build, both pre-existing:

1. core.c nii_otsu: -Waggressive-loop-optimizations ("iteration N invokes
   undefined behavior"). A false positive — nBin is guarded <=32767 and the sole
   caller passes 256, so the P/S index (row*nBin) never overflows int in practice.
   Compute the 2D index in size_t (new local OT2 macro) so no signed-overflow UB
   exists for gcc to reason about. Verified value-identical: all five -otsu modes
   produce byte-identical output before/after on a real image.

2. coreFLT.c overlay mismatch message: %lld vs int64_t (which is `long` on 64-bit
   Linux). Cast the args to (long long) to match the format on every platform.

+++ src/coreFLT.c
@@ -6513,7 +6513,7 @@ int main64(int argc, char *argv[]) {
 					goto fail;
 				}
 				if ((nim->nx != nim2->nx) || (nim->ny != nim2->ny) || (nim->nz != nim2->nz)) {
-					printfx("overlay dimensions do not match %lld×%lld×%lld != %lld×%lld×%lld \n", nim->nx, nim->ny, nim->nz, nim2->nx, nim2->ny, nim2->nz);
+					printfx("overlay dimensions do not match %lld×%lld×%lld != %lld×%lld×%lld \n", (long long)nim->nx, (long long)nim->ny, (long long)nim->nz, (long long)nim2->nx, (long long)nim2->ny, (long long)nim2->nz);
 					nifti_image_free(nim2);
 					goto fail;
 				}
